added all error handling logics

+++ src/schemas/schema.ts
@@ -2,11 +2,23 @@ const graphql = require('graphql');
 import Todo from '../models/todoSchema';
 import mongoose from 'mongoose';
 import _ from 'lodash';
+import { errorName } from './errors';
+const Joi = require('@hapi/joi');
 const { GraphQLObjectType, GraphQLBoolean, GraphQLString, GraphQLList, GraphQLInt, GraphQLSchema, GraphQLID } = graphql;
 
-const getError = (error: any): void => {
-    return error
-};
+// error handling through hapi/joi 
+const postApiParamsSchema = Joi.object({
+    title: Joi.string().required(),
+    description: Joi.string().required(),
+    isCompleted: Joi.boolean()
+});
+
+// error handling through hapi/joi 
+const putApiParamsSchema = Joi.object({
+    title: Joi.string(),
+    description: Joi.string(),
+    isCompleted: Joi.boolean()
+});
 
 const TodoType = new GraphQLObjectType({
     name: 'Todo',
@@ -15,8 +27,8 @@ const TodoType = new GraphQLObjectType({
         title: { type: GraphQLString },
         description: { type: GraphQLString },
         isCompleted: { type: GraphQLBoolean },
-        success: {type: GraphQLBoolean},
-        message: {type: GraphQLString}
+        success: { type: GraphQLBoolean },
+        message: { type: GraphQLString }
     })
 });
 
@@ -26,18 +38,31 @@ const RootQuery = new GraphQLObjectType({
         todo: {
             type: TodoType,
             args: { id: { type: GraphQLID } },
-            resolve(parent: any, args: any) {
+            async resolve(parent: any, args: any) {
                 // code to get data from database
-               
-                    return Todo.findById(args.id);
-               
+                const isValidId = mongoose.Types.ObjectId.isValid(args.id);
+                if (!isValidId) {
+                    throw new Error(errorName.INVALID_ID);
+                }
+                try {
+                    console.log("hello Biro");
+                    const todo = await Todo.findById(args.id);
+                    return todo;
+                } catch (error) {
+                    throw new Error(errorName.INTERNAL_SERVER_ERROR);
+                }
             }
         },
         todos: {
             type: new GraphQLList(TodoType),
-            resolve(parent: any, args: any) {
-                // code to get data from database
-                    return Todo.find({})
+            async  resolve(parent: any, args: any) {
+                try {
+                    //code to get data from database
+                    const todos = await Todo.find({});
+                    return todos;
+                } catch (error) {
+                    throw new Error(errorName.INTERNAL_SERVER_ERROR);
+                }
             }
         }
     }
@@ -54,12 +79,30 @@ const Mutation = new GraphQLObjectType({
                 isCompleted: { type: GraphQLBoolean }
             },
             resolve(parent: any, args: any) {
-                let todo = new Todo({
+                const { error } = postApiParamsSchema.validate({
                     title: args.title,
                     description: args.description,
                     isCompleted: args.isCompleted
                 });
-                return todo.save();
+                if (error) {
+                    const errName = error.details[0].context.label.toUpperCase() + "_ERROR";
+                    for (let i in errorName) {
+                        if (errorName[i] === errName) {
+                            throw new Error(errorName[i]);
+                        }
+                    }
+                }
+                try {
+                    let todo = new Todo({
+                        title: args.title,
+                        description: args.description,
+                        isCompleted: args.isCompleted
+                    });
+                    return todo.save();
+                } catch (error) {
+                    throw new Error(errorName.INTERNAL_SERVER_ERROR);
+                }
+
             }
         },
         deleteTodo: {
@@ -68,8 +111,17 @@ const Mutation = new GraphQLObjectType({
                 _id: { type: GraphQLID }
             },
             resolve(parent: any, args: any) {
-                let deletedTodo = Todo.findByIdAndDelete(args.id);
-                return deletedTodo;
+                const isValidId = mongoose.Types.ObjectId.isValid(args.id);
+                if (!isValidId) {
+                    throw new Error(errorName.INVALID_ID);
+                }
+                try {
+                    let deletedTodo = Todo.findByIdAndDelete(args.id);
+                    return deletedTodo;
+
+                } catch (error) {
+                    throw new Error(errorName.INTERNAL_SERVER_ERROR);
+                }
             }
         },
         updateTodo: {
@@ -80,28 +132,43 @@ const Mutation = new GraphQLObjectType({
                 description: { type: GraphQLString },
                 isCompleted: { type: GraphQLBoolean }
             },
-           async resolve(parent: any, args: any) {
+            async resolve(parent: any, args: any) {
                 let isValidId = mongoose.Types.ObjectId.isValid(args.id);
-                
                 if (!isValidId) {
-                    return {
-                        success: false,
-                        message: "Invalid todo Id"
-                    }
+                    throw new Error(errorName.INVALID_ID);
                 }
+
                 let reqBody: any = {
                     title: args.title,
                     isCompleted: args.isCompleted,
                     description: args.description
                 }
-                for (var propName in reqBody) { 
-                    if (reqBody[propName] === null || reqBody[propName] === undefined) {
-                      delete reqBody[propName];
+                const {error} = putApiParamsSchema.validate({
+                   ...reqBody
+                });
+                if (error) {
+                    const errName = error.details[0].context.label.toUpperCase() + "_ERROR";
+                    for (let i in errorName) {
+                        if (errorName[i] === errName) {
+                            throw new Error(errorName[i]);
+                        }
                     }
-                  }
-                return await Todo.findByIdAndUpdate(args.id, reqBody, {
-                    new: true
-                })
+                };
+                for (var propName in reqBody) {
+                    if (reqBody[propName] === null || reqBody[propName] === undefined) {
+                        delete reqBody[propName];
+                    }
+                }
+                if (Object.keys(reqBody).length < 1) {
+                    throw new Error(errorName.INVALID_BODY);
+                }
+                try {
+                    return await Todo.findByIdAndUpdate(args.id, reqBody, {
+                        new: true
+                    })
+                } catch (error) {
+                    throw new Error(errorName.INTERNAL_SERVER_ERROR);
+                }
             }
         }
     }
